zammad-ai/pyproject.toml
[project]
name = "zammad-ai"
license = "MIT"
version = "0.2.0"
description = "GenAI-powered agent for Zammad"
readme = "README.md"
requires-python = "==3.14.3"
authors = [
    { name = "Fabian Reinold", email = "fabian.reinold@muenchen.de" },
    { name = "Paul Ilitz", email = "paul.ilitz@muenchen.de" },
]
maintainers = [
    { name = "Fabian Reinold", email = "fabian.reinold@muenchen.de" },
]

dependencies = [
    "cryptography==46.0.6",
    "fastapi==0.135.1",
    "faststream[cli,kafka,prometheus]==0.6.7",
    "feedparser==6.0.12",
    "gradio==6.9.0",
    "httpx[socks]==0.28.1",
    "langchain-core==1.2.22",
    "langchain-openai==1.1.11",
    "langchain-qdrant==1.1.0",
    "langchain==1.2.12",
    "langfuse==4.0.1",
    "openai==2.28.0",
    "pandas==3.0.1",
    "pydantic-settings==2.13.1",
    "pydantic==2.12.5",
    "python-dotenv==1.2.2",
    "stamina==25.2.0",
    "truststore==0.10.4",
    "uvicorn==0.42.0",
]

[dependency-groups]
dev = [
    "ipykernel==7.2.0",
    "pre-commit==4.5.1",
    "pytest-asyncio==1.3.0",
    "pytest==9.0.2",
    "ruff==0.15.6",
    "ty==0.0.23",
]

[tool.uv]
add-bounds = "exact"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["test"]
asyncio_mode = "auto"
